fix(sharecode): type TokenID as uint16 and test new guards
#101

Match.TokenID was a uint32, so Encode's conversion to uint16 could
truncate any value above 0xFFFF. The share-code payload stores that
field in two bytes, so the Go type now matches the wire width and
steamresolve widens to uint32 only at the Game Coordinator and HTTP
boundary.

The capture job's invalid-PID path and the Ogg writer's negative
granule error each get a table test.

+++ internal/steamresolve/resolve.go
@@ -72,14 +72,14 @@ func (s *Service) Resolve(ctx context.Context, code string) (Result, error) {
 	if err != nil {
 		return Result{}, fmt.Errorf("%w: %v", ErrInvalidCode, err)
 	}
-	res := Result{MatchID: m.MatchID, OutcomeID: m.OutcomeID, TokenID: m.TokenID}
+	res := Result{MatchID: m.MatchID, OutcomeID: m.OutcomeID, TokenID: uint32(m.TokenID)}
 	if s == nil || s.transport == nil {
 		return res, nil
 	}
 	matches, err := s.transport.RequestMatch(ctx, steamgc.Request{
 		MatchID:   m.MatchID,
 		OutcomeID: m.OutcomeID,
-		Token:     m.TokenID,
+		Token:     uint32(m.TokenID),
 	})
 	if err != nil {
 		// Deliberately omits the share code and credentials from the message.
